Consolidate peripheral declarations.
