+++ firmware/application/portapack.hpp
@@ -1,0 +1,45 @@
+/*
+ * Copyright (C) 2014 Jared Boone, ShareBrained Technology, Inc.
+ *
+ * This file is part of PortaPack.
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2, or (at your option)
+ * any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; see the file COPYING.  If not, write to
+ * the Free Software Foundation, Inc., 51 Franklin Street,
+ * Boston, MA 02110-1301, USA.
+ */
+
+#include "portapack_io.hpp"
+
+#include "receiver_model.hpp"
+
+#include "spi_pp.hpp"
+#include "wm8731.hpp"
+
+#include "radio.hpp"
+
+namespace portapack {
+
+extern portapack::IO io;
+
+extern SPI ssp0;
+extern SPI ssp1;
+
+extern wolfson::wm8731::WM8731 audio_codec;
+
+extern ReceiverModel receiver_model;
+
+void init();
+void shutdown();
+
+} /* namespace portapack */
